Enhance GMTSAR sbas() wrapper to allow process detrended grids. Reorganize the functions between files.
