+++ pygmtsar/pygmtsar/SBAS_sbas.py
@@ -438,81 +438,6 @@ class SBAS_sbas(SBAS_sbas_gmtsar):
 
         return pd.DataFrame(data).sort_values(['ref_date', 'rep_date'])
 
-    #intf.tab format:   unwrap.grd  corr.grd  ref_id  rep_id  B_perp 
-    def intftab(self, baseline_pairs):
-        """
-        Generates a intf.tab formatted string based on the provided baseline pairs.
-
-        This function creates a string containing the interferogram metadata in the intf.tab format
-        for Sentinel-1 data. The string is generated based on the provided baseline_pairs DataFrame.
-
-        Parameters
-        ----------
-        baseline_pairs : pandas.DataFrame
-            A DataFrame containing the sorted list of baseline pairs with reference and repeat dates,
-            timelines, and baselines.
-
-        Returns
-        -------
-        str
-            A string containing the interferogram metadata in the intf.tab format.
-
-        """
-        import numpy as np
-        import datetime
-
-        # return a single subswath for None
-        subswath = self.get_subswath()
-
-        outs = []
-        for line in baseline_pairs.itertuples():
-            #print (line)
-            ref = line.ref_date.replace('-','')
-            jref = datetime.datetime.strptime(line.ref_date, '%Y-%m-%d').strftime('%Y%j')
-            rep = line.rep_date.replace('-','')
-            jrep = datetime.datetime.strptime(line.rep_date, '%Y-%m-%d').strftime('%Y%j')
-            bperp = np.round(line.rep_baseline - line.ref_baseline, 2)
-            outs.append(f'F{subswath}_{ref}_{rep}_unwrap.grd F{subswath}_{ref}_{rep}_corr.grd {jref} {jrep} {bperp}')
-        return '\n'.join(outs) + '\n'
-
-    def scenetab(self, baseline_pairs):
-        """
-        Generates a scene.tab formatted string based on the provided baseline pairs.
-
-        This function creates a string containing the scene metadata in the scene.tab format
-        for Sentinel-1 data. The string is generated based on the provided baseline_pairs DataFrame.
-
-        Parameters
-        ----------
-        baseline_pairs : pandas.DataFrame
-            A DataFrame containing the sorted list of baseline pairs with reference and repeat dates,
-            timelines, and baselines.
-
-        Returns
-        -------
-        str
-            A string containing the scene metadata in the scene.tab format.
-
-        """
-        import numpy as np
-        import datetime
-
-        mst = datetime.datetime.strptime(self.master, '%Y-%m-%d').strftime('%Y%j')
-        #print (self.master, mst)
-        outs = []
-        for line in baseline_pairs.itertuples():
-            #print (line)
-            ref = datetime.datetime.strptime(line.ref_date, '%Y-%m-%d').strftime('%Y%j')
-            yday_ref = np.round((line.ref_timeline - 2014)*365.25)
-            outs.append(f'{ref} {yday_ref}')
-            rep = datetime.datetime.strptime(line.rep_date, '%Y-%m-%d').strftime('%Y%j')
-            yday_rep = np.round((line.rep_timeline - 2014)*365.25)
-            outs.append(f'{rep} {yday_rep}')
-        outs = np.unique(outs)
-        return '\n'.join([out for out in outs if out.split(' ')[0]==mst]) + '\n' + \
-               '\n'.join([out for out in outs if out.split(' ')[0]!=mst]) + '\n'
-
-
     @staticmethod
     def triplets2pairs(triplets, pairs):
         """
