+++ pygmtsar/pygmtsar/SBAS_sbas_gmtsar.py
@@ -4,7 +4,7 @@ from .SBAS_detrend import SBAS_detrend
 
 class SBAS_sbas_gmtsar(SBAS_detrend):
 
-    def sbas(self, baseline_pairs, smooth=0, atm=0, debug=False):
+    def sbas(self, baseline_pairs, data='detrend', weight='corr', smooth=0, atm=0, debug=False):
         """
         Computes Small Baseline Subset (SBAS) InSAR time series analysis using the provided baseline pairs.
     
@@ -75,7 +75,7 @@ class SBAS_sbas_gmtsar(SBAS_detrend):
         rng_pixel_size = 300000000 / rng_samp_rate / 2
         rng = np.round(rng_pixel_size * (xmin+xmax) /2 + near_range)
 
-        intf_tab = self.intftab(baseline_pairs)
+        intf_tab = self.intftab(baseline_pairs, data=data, weight=weight)
         pipe1 = os.pipe()
         os.write(pipe1[1], bytearray(intf_tab, 'ascii'))
         os.close(pipe1[1])
@@ -116,3 +116,77 @@ class SBAS_sbas_gmtsar(SBAS_detrend):
             #print (jdate, date)
 
         return
+
+    #intf.tab format:   unwrap.grd  corr.grd  ref_id  rep_id  B_perp 
+    def intftab(self, baseline_pairs, data='unwrap', weight='corr'):
+        """
+        Generates a intf.tab formatted string based on the provided baseline pairs.
+
+        This function creates a string containing the interferogram metadata in the intf.tab format
+        for Sentinel-1 data. The string is generated based on the provided baseline_pairs DataFrame.
+
+        Parameters
+        ----------
+        baseline_pairs : pandas.DataFrame
+            A DataFrame containing the sorted list of baseline pairs with reference and repeat dates,
+            timelines, and baselines.
+
+        Returns
+        -------
+        str
+            A string containing the interferogram metadata in the intf.tab format.
+
+        """
+        import numpy as np
+        import datetime
+
+        # return a single subswath for None
+        subswath = self.get_subswath()
+
+        outs = []
+        for line in baseline_pairs.itertuples():
+            #print (line)
+            ref = line.ref_date.replace('-','')
+            jref = datetime.datetime.strptime(line.ref_date, '%Y-%m-%d').strftime('%Y%j')
+            rep = line.rep_date.replace('-','')
+            jrep = datetime.datetime.strptime(line.rep_date, '%Y-%m-%d').strftime('%Y%j')
+            bperp = np.round(line.rep_baseline - line.ref_baseline, 2)
+            outs.append(f'F{subswath}_{ref}_{rep}_{data}.grd F{subswath}_{ref}_{rep}_{weight}.grd {jref} {jrep} {bperp}')
+        return '\n'.join(outs) + '\n'
+
+    def scenetab(self, baseline_pairs):
+        """
+        Generates a scene.tab formatted string based on the provided baseline pairs.
+
+        This function creates a string containing the scene metadata in the scene.tab format
+        for Sentinel-1 data. The string is generated based on the provided baseline_pairs DataFrame.
+
+        Parameters
+        ----------
+        baseline_pairs : pandas.DataFrame
+            A DataFrame containing the sorted list of baseline pairs with reference and repeat dates,
+            timelines, and baselines.
+
+        Returns
+        -------
+        str
+            A string containing the scene metadata in the scene.tab format.
+
+        """
+        import numpy as np
+        import datetime
+
+        mst = datetime.datetime.strptime(self.master, '%Y-%m-%d').strftime('%Y%j')
+        #print (self.master, mst)
+        outs = []
+        for line in baseline_pairs.itertuples():
+            #print (line)
+            ref = datetime.datetime.strptime(line.ref_date, '%Y-%m-%d').strftime('%Y%j')
+            yday_ref = np.round((line.ref_timeline - 2014)*365.25)
+            outs.append(f'{ref} {yday_ref}')
+            rep = datetime.datetime.strptime(line.rep_date, '%Y-%m-%d').strftime('%Y%j')
+            yday_rep = np.round((line.rep_timeline - 2014)*365.25)
+            outs.append(f'{rep} {yday_rep}')
+        outs = np.unique(outs)
+        return '\n'.join([out for out in outs if out.split(' ')[0]==mst]) + '\n' + \
+               '\n'.join([out for out in outs if out.split(' ')[0]!=mst]) + '\n'
